remove field.length so that the length is only kept in field.typeval.length
To ensure that exactly x bit are assigned to ArrayValues or OpaqueValues if the calculated length
is x a check is performed during assignment of the value. The length expression that was formerly
kept in field.length must now be set to the CompositValue using the set_expected_size method prior
to assigning a value. If the CompositValue does not have a length it must not be set.