add refinements
All Refinements that are possibly used in this MessageValues are added to it in pyrflx.py. This includes Refinements
that are possibly used in nested messages i.e. refinements for ipv4 are also added to an ethernet message
(if Ethernet -> Ipv4 -> UDP refinements are defined). If an OpaqueValue is assigned using the set() method in
MessageValue a check is performed whether applicable refinements exist. Only then it is passed
to the OpaqueValue before assigning the bytes or Bitstring value. The list of all refinements known to the Top-Level
Message is passed to the OpaqueValue to ensure that further nested messages can be created recursively.
Ref #199